feat: implement local instance database bootstrap scripts
- Add database-bootstrap.ts for recreating the database and importing schema.
- Introduce docker-psql.ts for executing PostgreSQL commands in Docker containers.
- Create keycloak-sync.ts to fetch and sync Keycloak users.
- Implement logging.ts for standardized logging steps.
- Add parse-options.ts for command-line argument parsing.
- Create source-instance.ts to load metadata from the source instance.
- Implement summary.ts to summarize the target state of the database.
- Add tests for instance registry commands and context.
- Create command-context.ts for managing database connections and command execution.
- Implement formatters.ts for rendering command results.
- Add mutation-commands.ts for handling instance registry mutations.
- Create parse-options.ts for parsing CLI options for instance registry commands.
- Implement read-commands.ts for executing read commands in the instance registry.
- Add shared.ts for common types and utility functions related to instance registry commands.
