+++ package.json
@@ -31,13 +31,16 @@
     "test:integration": "env -u NO_COLOR nx run-many -t test:integration",
     "test:e2e": "env -u NO_COLOR nx run sva-studio-react:test:e2e",
     "test:acceptance:iam": "env -u NO_COLOR nx run sva-studio-react:test:acceptance",
-    "test:ci": "pnpm clean:generated-source-artifacts && pnpm check:file-placement && pnpm check:plugin-ui-boundary && pnpm check:openapi:iam && pnpm test:types && pnpm test:unit && pnpm test:coverage && pnpm complexity-gate && pnpm test:ops:waste-calendar-example-pdf",
+    "test:ci": "pnpm clean:generated-source-artifacts && pnpm check:file-placement && pnpm check:plugin-ui-boundary && pnpm check:openapi:iam && pnpm test:types && pnpm test:unit && pnpm test:coverage && pnpm complexity-gate && pnpm test:ops:critical",
     "coverage-gate": "tsx scripts/ci/coverage-gate.ts",
     "patch-coverage-gate": "tsx scripts/ci/patch-coverage-gate.ts",
     "sonar-new-code-gate": "tsx scripts/ci/sonar-new-code-gate.ts",
     "sonar:prepare-lcov": "tsx scripts/ci/prepare-sonar-lcov.ts",
     "complexity-gate": "tsx scripts/ci/complexity-gate.ts",
+    "test:ops:instance-registry": "pnpm exec vitest run scripts/ops/instance-registry.test.ts",
+    "test:ops:bootstrap-local-instance-db": "pnpm exec vitest run scripts/ops/bootstrap-local-instance-db.test.ts",
     "test:ops:waste-calendar-example-pdf": "node --import tsx --test scripts/ops/waste-calendar-example-pdf.test.ts",
+    "test:ops:critical": "pnpm test:ops:instance-registry && pnpm test:ops:bootstrap-local-instance-db && pnpm test:ops:waste-calendar-example-pdf",
     "clean:generated-source-artifacts": "tsx scripts/ci/clean-generated-source-artifacts.ts",
     "clean:source-artifacts": "pnpm clean:generated-source-artifacts",
     "check:file-placement": "tsx scripts/ci/check-file-placement.ts",
