+++ packages/plugin-events/tests/vitest-config.test.ts
@@ -1,0 +1,8 @@
+import vitestConfig from '../vitest.config';
+import { sharedCoverageConfig } from '../../../vitest.config';
+
+describe('plugin-events vitest coverage config', () => {
+  it('inherits the shared coverage reporters required by the CI gate', () => {
+    expect(vitestConfig.test?.coverage).toMatchObject(sharedCoverageConfig);
+  });
+});
