+++ packages/plugin-poi/vitest.config.ts
@@ -1,6 +1,7 @@
 import { dirname, resolve } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { defineConfig } from 'vitest/config';
+import { sharedCoverageConfig } from '../../vitest.config';
 
 const currentDir = dirname(fileURLToPath(import.meta.url));
 
@@ -18,5 +19,6 @@ export default defineConfig({
     globals: true,
     name: 'plugin-poi',
     setupFiles: [],
+    coverage: sharedCoverageConfig,
   },
 });
