+++ scripts/ops/bootstrap-local-instance-db/logging.ts
@@ -1,0 +1,5 @@
+export type LogStep = (message: string) => void;
+
+export const logStep: LogStep = (message) => {
+  process.stdout.write(`\n==> ${message}\n`);
+};
